applications/slackbot_chatgpt/src/requirements.txt
functions-framework==3.3.0
openai == 0.26.4
slack-bolt == 1.16.1
google-cloud-logging == 3.5.0